requirements.txt
requests==2.25.1
paho-mqtt==1.5.1
pytz==2021.1
python-dotenv==0.15.0